pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tabula-writer"
version = "2.0.0"
authors = [
  { name="Pendragon", email="author@example.com" },
]
description = "A markdown-based writing application with multiple panels"
requires-python = ">=3.8"
dependencies = [
    "PyQt6",
    "python-docx",
    "fpdf2",
    "pyspellchecker",
]

[project.gui-scripts]
tabula-writer = "tabula_writer.main_qt:main"
